perf: restructure imports
