+++ packages/perf-benchmarks/best.server.config.js
@@ -4,37 +4,13 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
+
+const config = require('./best.base.config');
+
 module.exports = {
+    ...config,
+
     projectName: '@lwc/engine-server',
     testMatch: ['**/__benchmarks__/engine-server/**/*.benchmark.js'],
-    plugins: [
-        '<rootDir>/best-plugins/server-resolver.js',
-        [
-            '@lwc/rollup-plugin',
-            {
-                rootDir: '<rootDir>/src/',
-                exclude: ['/engine.js$/', '/@best/runtime/'],
-            },
-        ],
-        ['@rollup/plugin-replace', { 'process.env.NODE_ENV': JSON.stringify('production') }],
-    ],
-    benchmarkOnClient: false,
-    benchmarkIterations: 60,
-    specs: { name: 'chrome.headless', version: '80' },
-    runners: [
-        {
-            alias: 'default',
-            runner: '@best/runner-headless',
-        },
-        {
-            alias: 'remote',
-            runner: '@best/runner-remote',
-            config: {
-                uri: process.env.HUB_URI,
-                options: {
-                    authToken: process.env.HUB_AUTH_TOKEN,
-                },
-            },
-        },
-    ],
+    plugins: ['<rootDir>/best-plugins/server-resolver.js', ...config.plugins],
 };
