+++ packages/perf-benchmarks/src/__benchmarks__/engine-dom/benchmark-table-wc/wc-update-10th-1k.benchmark.js
@@ -6,9 +6,8 @@
  */
 import Table from 'benchmark/tableComponent';
 import Row from 'benchmark/tableComponentRow';
-
-import Store from '../../../tableStore';
-import { insertTableComponent, destroyTableComponent } from '../../../utils';
+import Store from 'benchmark/store';
+import { insertTableComponent, destroyTableComponent } from 'benchmark/utils/utils';
 
 customElements.define('benchmark-table-component', Table.CustomElementConstructor);
 // the row can be optionally defined, but this benchmark always do it so we know how costly it is.
