+++ packages/perf-benchmarks/src/__benchmarks__/engine-dom/benchmark-table/table-clear-1k.benchmark.js
@@ -6,10 +6,10 @@
  */
 
 import { createElement } from 'lwc';
-import Table from 'benchmark/table';
 
-import Store from '../../../tableStore';
-import { insertTableComponent, destroyTableComponent } from '../../../utils';
+import Table from 'benchmark/table';
+import Store from 'benchmark/store';
+import { insertTableComponent, destroyTableComponent } from 'benchmark/utils';
 
 benchmark(`benchmark-table/clear/1k`, () => {
     let tableElement;
